Added try-catch blocks to fix race conditions
Issue 1: App crashes if user keeps pressing the answer button very fast
or shakes it incessantly even when the animation is seemingly going on.

Cause: Even though the app checks if the animation is completed before
asking for a repeat - the resource not found exception is caused by the
tie between the end of an animation and the time needed to load a new
image for the animation.

Proposed Solution: Try.. catch will not cause a crash but will not solve
it either. There needs to be a duration set between each calls. (To be
investigated)